requirements.txt
django==1.11.13
djangorestframework==3.8.2
psycopg2==2.7.4
requests